reorgonize signal code structure
