+++ includes/signals/signals.hpp
@@ -1,0 +1,14 @@
+#ifndef SIGNALS_HPP
+#define SIGNALS_HPP
+
+#include <csignal>
+#include <iostream>
+
+extern volatile sig_atomic_t g_sig;
+
+void sigIntHandler(int sig);
+void setupSigHandler();
+
+
+
+#endif
